줌인 줌아웃시 한번만 줌을 할시에 그 다음부터는 줌 할 때 시간이 오래 걸리지 않음

+++ src/main/java/com/freepass/dto/ConstructionDTO.java
@@ -11,14 +11,14 @@ public class ConstructionDTO {
     private double latitude; // GPS 위도
     private double longitude; // GPS 경도
 
-    public ConstructionDTO(String name, String contractor, String location, String startDate, String endDate, int x, int y) {
+    public ConstructionDTO(String name, String contractor, String location, String startDate, String endDate, double x, double y) {
         this.name = name;
         this.contractor = contractor;
         this.location = location;
         this.startDate = startDate;
         this.endDate = endDate;
-        this.x = x;
-        this.y = y;
+        this.x = (int) x;
+        this.y = (int) y;
     }
 
     // 복사 생성자 (안전한 객체 복사를 위해)
